Jogosultságok hozzáadása

+++ Controllers/HR/DepartmentController.cs
@@ -17,17 +17,27 @@ public class DepartmentController : ControllerBase
         _service = service;
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] DepartmentDto dto)
     {
+
+        if (!User.HasClaim("permission", "CREATE:DEPARTMENTS"))
+            return Forbid();
+
         var result = await _service.CreateAsync(dto);
 
         return Ok(result);
     }
 
+    [Authorize]
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
+
+        if (!User.HasClaim("permission", "VIEW:DEPARTMENTS"))
+            return Forbid();
+
         var result = await _service.GetAllAsync();
 
         if (result == null) NotFound();
@@ -35,9 +45,14 @@ public class DepartmentController : ControllerBase
         return Ok(result);
     }
 
+    [Authorize]
     [HttpGet("{id}")]
     public async Task<IActionResult> GetById(int id)
     {
+
+        if (!User.HasClaim("permission", "VIEW:DEPARTMENTS"))
+            return Forbid();
+
         var result = await _service.GetByIdAsync(id);
 
         if (result == null) NotFound();
@@ -45,9 +60,14 @@ public class DepartmentController : ControllerBase
         return Ok(result);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<IActionResult> Update(int id, [FromBody] DepartmentDto dto)
     {
+
+        if (!User.HasClaim("permission", "EDIT:DEPARTMENTS"))
+            return Forbid();
+
         var result = await _service.UpdateAsync(id, dto);
 
         if (!result) NotFound();
@@ -55,9 +75,14 @@ public class DepartmentController : ControllerBase
         return Ok(result);
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<IActionResult> Delete(int id)
     {
+
+        if (!User.HasClaim("permission", "DELETE:DEPARTMENTS"))
+            return Forbid();
+
         var result = await _service.DeleteAsync(id);
 
         if (!result) NotFound();
