+++ Controllers/HR/PositionController.cs
@@ -18,17 +18,27 @@ public class PositionController : ControllerBase
         _service = service;
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] PositionDto dto)
     {
+
+        if (!User.HasClaim("permission", "CREATE:POSITIONS"))
+            return Forbid();
+
         var result = await _service.CreateAsync(dto);
 
         return Ok(result);
     }
 
+    [Authorize]
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
+
+        if (!User.HasClaim("permission", "VIEW:POSITIONS"))
+            return Forbid();
+
         var result = await _service.GetAllAsync();
 
         if (result == null) NotFound();
@@ -36,9 +46,14 @@ public class PositionController : ControllerBase
         return Ok(result);
     }
 
+    [Authorize]
     [HttpGet("{id}")]
     public async Task<IActionResult> GetById(int id)
     {
+
+        if (!User.HasClaim("permission", "VIEW:POSITIONS"))
+            return Forbid();
+
         var result = await _service.GetByIdAsync(id);
 
         if (result == null) NotFound();
@@ -46,9 +61,14 @@ public class PositionController : ControllerBase
         return Ok(result);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<IActionResult> Update(int id, [FromBody] PositionDto dto)
     {
+
+        if (!User.HasClaim("permission", "EDIT:POSITIONS"))
+            return Forbid();
+
         var result = await _service.UpdateAsync(id, dto);
 
         if (!result) NotFound();
@@ -56,9 +76,14 @@ public class PositionController : ControllerBase
         return Ok(result);
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<IActionResult> Delete(int id)
     {
+
+        if (!User.HasClaim("permission", "DELETE:POSITIONS"))
+            return Forbid();
+
         var result = await _service.DeleteAsync(id);
 
         if (!result) NotFound();
