+++ Controllers/Procurement/GoodsReceiptController.cs
@@ -17,18 +17,28 @@ public class GoodsReceiptController : ControllerBase
         _service = service;
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] CreateGoodsReceiptDto dto)
     {
+
+        if (!User.HasClaim("permission", "CREATE:GOODS_RECEIPTS"))
+            return Forbid();
+
         var result = await _service.CreateAsync(dto);
 
         return Ok(result);
     }
 
+    [Authorize]
     [HttpGet]
-    public async Task<IActionResult> GetAll()
+    public async Task<IActionResult> GetAll([FromQuery] int? purchaseOrderId, [FromQuery] int? warehouseId, [FromQuery] DateTime? date)
     {
-        var result = await _service.GetAllAsync();
+
+        if (!User.HasClaim("permission", "VIEW:GOODS_RECEIPTS"))
+            return Forbid();
+
+        var result = await _service.GetAllAsync(purchaseOrderId, warehouseId, date);
 
         if (result == null)
             return NotFound();
@@ -36,10 +46,14 @@ public class GoodsReceiptController : ControllerBase
         return Ok(result);
     }
 
-    
+    [Authorize]
     [HttpGet("{id}")]
     public async Task<IActionResult> GetById(int id)
     {
+
+        if (!User.HasClaim("permission", "VIEW:GOODS_RECEIPTS"))
+            return Forbid();
+
         var result = await _service.GetByIdAsync(id);
 
         if (result == null)
@@ -48,9 +62,14 @@ public class GoodsReceiptController : ControllerBase
         return Ok(result);
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<IActionResult> Delete(int id)
     {
+
+        if (!User.HasClaim("permission", "DELETE:GOODS_RECEIPTS"))
+            return Forbid();
+
         var result = await _service.DeleteAsync(id);
 
         if (!result)
