+++ Controllers/Procurement/SupplierController.cs
@@ -17,17 +17,27 @@ public class SupplierController : ControllerBase
         _service = service;
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] CreateSupplierDto dto)
     {
+
+        if (!User.HasClaim("permission", "CREATE:SUPPLIERS"))
+            return Forbid();
+
         var result = await _service.CreateAsync(dto);
 
         return Ok(result);
     }
 
+    [Authorize]
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
+
+        if (!User.HasClaim("permission", "VIEW:SUPPLIERS"))
+            return Forbid();
+
         var result = await _service.GetAllAsync();
 
         if (result == null)
@@ -36,9 +46,14 @@ public class SupplierController : ControllerBase
         return Ok(result);
     }
 
+    [Authorize]
     [HttpGet("{id}")]
     public async Task<IActionResult> GetById(int id)
     {
+
+        if (!User.HasClaim("permission", "VIEW:SUPPLIERS"))
+            return Forbid();
+
         var result = await _service.GetByIdAsync(id);
 
         if (result == null)
@@ -47,9 +62,14 @@ public class SupplierController : ControllerBase
         return Ok(result);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<IActionResult> Update(int id, [FromBody] UpdateSupplierDto dto)
     {
+
+        if (!User.HasClaim("permission", "VIEW:SUPPLIERS"))
+            return Forbid();
+
         var result = await _service.UpdateAsync(id, dto);
 
         if (!result)
@@ -58,9 +78,14 @@ public class SupplierController : ControllerBase
         return Ok(result);
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<IActionResult> Delete(int id)
     {
+
+        if (!User.HasClaim("permission", "DELETE:SUPPLIERS"))
+            return Forbid();
+
         var result = await _service.DeleteAsync(id);
 
         if (!result)
