+++ Controllers/ProductController.cs
@@ -18,17 +18,27 @@ public class ProductController : ControllerBase
         _service = service;
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] CreateProductDto dto)
     {
+
+        if (!User.HasClaim("permission", "CREATE:PRODUCTS"))
+            return Forbid();
+
         var result = await _service.CreateAsync(dto);
 
         return Ok(result);
     }
 
+    [Authorize]
     [HttpGet]
     public async Task<IActionResult> GetAll([FromQuery] string? sku, [FromQuery] string? name, [FromQuery] bool? active)
     {
+
+        if (!User.HasClaim("permission", "VIEW:PRODUCTS"))
+            return Forbid();
+
         var result = await _service.GetAllAsync(sku, name, active);
 
         if (result == null)
@@ -37,9 +47,14 @@ public class ProductController : ControllerBase
         return Ok(result);
     }
 
+    [Authorize]
     [HttpGet("{id}")]
     public async Task<IActionResult> GetById(int id)
     {
+
+        if (!User.HasClaim("permission", "VIEW:PRODUCTS"))
+            return Forbid();
+
         var result = await _service.GetByIdAsync(id);
 
         if (result == null)
@@ -48,9 +63,14 @@ public class ProductController : ControllerBase
         return Ok(result);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<IActionResult> Update(int id, [FromBody] Product model)
     {
+
+        if (!User.HasClaim("permission", "EDIT:PRODUCTS"))
+            return Forbid();
+
         var result = await _service.UpdateAsync(id, model);
 
         if (!result)
@@ -59,9 +79,14 @@ public class ProductController : ControllerBase
         return Ok(result);
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<IActionResult> Delete(int id)
     {
+
+        if (!User.HasClaim("permission", "DELETE:PRODUCTS"))
+            return Forbid();
+
         var result = await _service.DeleteAsync(id);
 
         if (!result)
